Wait to add shadows to fix infity error
If we add shadow immediatly after changing the image src, the natural
height is not recognized, and it causes an error where the shadow
height is calculated as 'infinity', and the shadows are not added.
This means that the user always had to return the first time, and
load the image again.